feat(fearGreed): extract FearGreedGauge primitive + CNN-style mini gauges
